+++ deno.json
@@ -21,13 +21,14 @@
       "@/lib/core/design": "./src/lib/core/design.ts",
       "@/lib/core/dom": "./src/lib/core/dom.ts",
       "@/lib/core/elements": "./src/lib/core/elements.ts",
+      "@/lib/core/markdown": "./src/lib/core/markdown.ts",
       "@/lib/core/properties": "./src/lib/core/properties.ts",
       "@/lib/core/sanityze": "./src/lib/core/sanityze.ts",
       "@/lib/core/store": "./src/lib/core/store.ts",
       "@/lib/core/style": "./src/lib/core/style.ts",
       "@/lib/core/stylerules": "./src/lib/core/stylerules.ts",
-      "@/lib/core/types": "./src/lib/core/types.ts",
       
+      "@/lib/server/inserter": "./src/lib/server/inserter.ts",
       "@/lib/server/markdown": "./src/lib/server/markdown.ts",
       "@/lib/server/stringify": "./src/lib/server/stringify.ts",
       "@/lib/server/util": "./src/lib/server/util.ts",
