+++ src/cli/html.ts
@@ -1,7 +1,7 @@
 import type { Attribute, HNode, HRootPageFn } from "@/lib/core/component";
 import { DOCTYPE } from "@/lib/core/elements";
 import type { HComponentInsert, Store } from "@/lib/core/store";
-import { insertNodes } from "@/lib/core/style";
+import { insertNodes } from "@/lib/server/inserter";
 import { stringifyToHtml } from "@/lib/server/stringify";
 
 export async function bundleHtml(
