+++ src/lib/core/style.ts
@@ -1,7 +1,5 @@
-import type { Attribute, HAnyComponentFn, HElement, HNode } from "@/lib/core/component";
+import type { HAnyComponentFn } from "@/lib/core/component";
 import type { Properties } from "@/lib/core/properties";
-import { property_names } from "@/lib/core/properties";
-import type { HComponent } from "@/lib/core/store";
 import { unionArrayOfRecords } from "@/lib/core/util";
 
 export type PropertyOf<T extends keyof Properties> = Properties[T];
@@ -45,97 +43,7 @@ export function atStyle(
         });
 }
 
-// Inserter (HNode).
-export function insertNodes(root: HNode, selector: Selector[], insert: HNode[], search_deep: boolean): HNode {
-    if (typeof root === "string") {
-        return root;
-    }
-
-    if (selector.length === 0) {
-        return root;
-    }
-
-    if (isCombinator(selector[0])) {
-        switch (selector[0]) {
-            case ">":
-                return insertNodes(root, selector.slice(1), insert, false);
-            default:
-                throw new Error("insertElementsCombinator: unsupported combinator.");
-        }
-    }
-
-    if (matchCompoundSelector(normalizeSelector(selector[0]), root)) {
-        const child =
-            selector.length === 1
-                ? [...root.child, ...insert]
-                : root.child.map((c) => insertNodes(c, selector.slice(1), insert, true));
-        return {
-            tag: root.tag,
-            attribute: root.attribute,
-            child,
-        };
-    }
-
-    if (search_deep) {
-        return {
-            tag: root.tag,
-            attribute: root.attribute,
-            child: root.child.map((c) => insertNodes(c, selector, insert, true)),
-        };
-    }
-    return root;
-}
-
-function hasClass(className: string, attribute: Attribute): boolean {
-    if (attribute.class === undefined || typeof attribute.class !== "string") {
-        return false;
-    }
-    return attribute.class.includes(className);
-}
-
-function matchCompoundSelector(selector: CompoundSelector, element: HElement<{ id?: string }>): boolean {
-    for (const s of selector) {
-        if (typeof s !== "string") {
-            throw new Error("matchCompoundSelector: ComponentFn is not supported.");
-        }
-        if (s.startsWith(".")) {
-            if (!hasClass(s.slice(1), element.attribute)) {
-                return false;
-            }
-        } else if (s.startsWith("#")) {
-            if (element.attribute.id === undefined || element.attribute.id !== s.slice(1)) {
-                return false;
-            }
-        } else if (s !== "*") {
-            if (element.tag !== s) {
-                return false;
-            }
-        }
-    }
-    return true;
-}
-
-function isCompoundSelector(s: Selector): s is CompoundSelector {
-    return Array.isArray(s);
-}
-
-function isCombinator(s: Selector): s is Combinator {
-    return typeof s === "string" && (s === " " || s === ">" || s === "+" || s === "~");
-}
-
-function isSimpleSelector(s: Selector): s is SimpleSelector {
-    return (typeof s === "string" && !isCombinator(s)) || typeof s === "function";
-}
-
-function isString(selector: SimpleSelector): selector is string {
-    return typeof selector === "string";
-}
-
-function isComponentFn(selector: SimpleSelector): selector is HAnyComponentFn {
-    return typeof selector === "function";
-}
-
-function normalizeSelector(context: Selector): CompoundSelector {
+export function normalizeSelector(context: Selector): CompoundSelector {
     if (isCombinator(context)) {
         return [context];
     }
@@ -153,93 +61,22 @@ function normalizeSelector(context: Selector): CompoundSelector {
     throw new Error(`createSelector: internal error. type mismatch 1 at ${context}`);
 }
 
-// Stringify
-export function stringifyToCss(components: HComponent[]): string {
-    return `@charset "utf-8";\n@layer base, font, low, main, high;\n${components.map(rulesToString).join("")}`;
+export function isCompoundSelector(s: Selector): s is CompoundSelector {
+    return Array.isArray(s);
 }
 
-export function rulesToString(element: HComponent): string {
-    const res: string[] = [];
-    for (const rule of element.style) {
-        const selectors_string = rule.selector.map(selectorContextToString).join(", ");
-        const propaties_string = propertiesToString(rule.properties);
-        if (rule.atrules) {
-            const has_layer = rule.atrules.some((x) => x.length > 0 && x[0].localeCompare("@layer") === 0);
-            const atrules = has_layer ? rule.atrules : [["@layer", "main"], ...rule.atrules];
-            res.push(
-                `${atrules.map((x) => x.join(" ")).join(" { ")} { ${selectors_string} { ${propaties_string} } ${" } ".repeat(atrules.length)}\n`,
-            );
-        } else {
-            res.push(`@layer main { ${selectors_string} { ${propaties_string} } }\n`);
-        }
-    }
-
-    return res.join("");
+export function isCombinator(s: Selector): s is Combinator {
+    return typeof s === "string" && (s === " " || s === ">" || s === "+" || s === "~");
 }
 
-function selectorContextToString(selector: SelectorContext): string {
-    if (isSelector(selector)) {
-        return selector.map(selectorToString).join(" ");
-    }
-    return selectorToString(selector);
+function isSimpleSelector(s: Selector): s is SimpleSelector {
+    return (typeof s === "string" && !isCombinator(s)) || typeof s === "function";
 }
 
-function selectorToString(selector: Selector): string {
-    if (isCompoundSelector(selector)) {
-        return selector.map(normalizeSelector).join("");
-    }
-    return normalizeSelector(selector).join("");
+function isString(selector: SimpleSelector): selector is string {
+    return typeof selector === "string";
 }
 
-function isSelector(context: SelectorContext): context is Selector[] {
-    return Array.isArray(context);
-}
-
-function valueToString(value: string | string[] | string[][]): string {
-    if (typeof value === "string") {
-        return singleValueToString(value);
-    }
-    if (value.every((x) => typeof x === "string")) {
-        return spaceSeparatedValueToString(value);
-    }
-    return commaSeparatedValueToString(value);
-}
-
-function singleValueToString(value: string): string {
-    if (value.length > 512) {
-        throw new Error(`propertyesToString: value length must be under 512 characters. (${value})`);
-    }
-    return value;
-}
-
-function spaceSeparatedValueToString(value: string[]): string {
-    return value.map(singleValueToString).join(" ");
-}
-
-function commaSeparatedValueToString(value: string[][]): string {
-    return value
-        .map((x) => (typeof x === "string" ? singleValueToString(x) : spaceSeparatedValueToString(x)))
-        .join(", ");
-}
-
-function propertiesToString(properties: Properties): string {
-    return Object.entries(properties)
-        .map(([raw_key, raw_value]) => {
-            if (!validatePropertyName(raw_key)) {
-                throw new Error(`propertyesToString: invalid key. (${raw_key})})`);
-            }
-            if (raw_value === undefined) {
-                throw new Error(`propertiesToString: undefined value is not allowed in property ${raw_key}`);
-            }
-            const key = raw_key.replaceAll("_", "-");
-            const value = valueToString(raw_value);
-            return `${key}: ${value};`;
-        })
-        .join(" ");
-}
-
-const allowed_property_names = /* @__PURE__*/ new Set<string>(property_names);
-
-function validatePropertyName(name: string): boolean {
-    return allowed_property_names.has(name);
+function isComponentFn(selector: SimpleSelector): selector is HAnyComponentFn {
+    return typeof selector === "function";
 }
